+++ pygmt/helpers/utils.py
@@ -12,11 +12,33 @@ import sys
 import time
 import webbrowser
 from collections.abc import Iterable, Mapping, Sequence
+from pathlib import Path
 from typing import Any, Literal
 
 import xarray as xr
 from pygmt.encodings import charset
 from pygmt.exceptions import GMTInvalidInput
+
+# Type hints for the list of encodings supported by PyGMT.
+Encoding = Literal[
+    "ascii",
+    "ISOLatin1+",
+    "ISO-8859-1",
+    "ISO-8859-2",
+    "ISO-8859-3",
+    "ISO-8859-4",
+    "ISO-8859-5",
+    "ISO-8859-6",
+    "ISO-8859-7",
+    "ISO-8859-8",
+    "ISO-8859-9",
+    "ISO-8859-10",
+    "ISO-8859-11",
+    "ISO-8859-13",
+    "ISO-8859-14",
+    "ISO-8859-15",
+    "ISO-8859-16",
+]
 
 
 def _validate_data_input(
@@ -58,7 +80,7 @@ def _validate_data_input(
     >>> _validate_data_input(
     ...     data=pd.DataFrame(data, columns=["x", "y"]),
     ...     required_z=True,
-    ...     kind="matrix",
+    ...     kind="vectors",
     ... )
     Traceback (most recent call last):
         ...
@@ -66,7 +88,7 @@ def _validate_data_input(
     >>> _validate_data_input(
     ...     data=xr.Dataset(pd.DataFrame(data, columns=["x", "y"])),
     ...     required_z=True,
-    ...     kind="matrix",
+    ...     kind="vectors",
     ... )
     Traceback (most recent call last):
         ...
@@ -96,46 +118,34 @@ def _validate_data_input(
     if data is None:  # data is None
         if x is None and y is None:  # both x and y are None
             if required_data:  # data is not optional
-                raise GMTInvalidInput("No input data provided.")
+                msg = "No input data provided."
+                raise GMTInvalidInput(msg)
         elif x is None or y is None:  # either x or y is None
-            raise GMTInvalidInput("Must provide both x and y.")
+            msg = "Must provide both x and y."
+            raise GMTInvalidInput(msg)
         if required_z and z is None:  # both x and y are not None, now check z
-            raise GMTInvalidInput("Must provide x, y, and z.")
+            msg = "Must provide x, y, and z."
+            raise GMTInvalidInput(msg)
     else:  # data is not None
         if x is not None or y is not None or z is not None:
-            raise GMTInvalidInput("Too much data. Use either data or x/y/z.")
-        # For 'matrix' kind, check if data has the required z column
-        if kind == "matrix" and required_z:
-            if hasattr(data, "shape"):  # np.ndarray or pd.DataFrame
-                if len(data.shape) == 1 and data.shape[0] < 3:
-                    raise GMTInvalidInput("data must provide x, y, and z columns.")
-                if len(data.shape) > 1 and data.shape[1] < 3:
-                    raise GMTInvalidInput("data must provide x, y, and z columns.")
-            if hasattr(data, "data_vars") and len(data.data_vars) < 3:  # xr.Dataset
-                raise GMTInvalidInput("data must provide x, y, and z columns.")
+            msg = "Too much data. Use either data or x/y/z."
+            raise GMTInvalidInput(msg)
+        # check if data has the required z column
+        if required_z:
+            msg = "data must provide x, y, and z columns."
+            if kind == "matrix" and data.shape[1] < 3:
+                raise GMTInvalidInput(msg)
+            if kind == "vectors":
+                if hasattr(data, "shape") and (
+                    (len(data.shape) == 1 and data.shape[0] < 3)
+                    or (len(data.shape) > 1 and data.shape[1] < 3)
+                ):  # np.ndarray or pd.DataFrame
+                    raise GMTInvalidInput(msg)
+                if hasattr(data, "data_vars") and len(data.data_vars) < 3:  # xr.Dataset
+                    raise GMTInvalidInput(msg)
 
 
-def _check_encoding(
-    argstr: str,
-) -> Literal[
-    "ascii",
-    "ISOLatin1+",
-    "ISO-8859-1",
-    "ISO-8859-2",
-    "ISO-8859-3",
-    "ISO-8859-4",
-    "ISO-8859-5",
-    "ISO-8859-6",
-    "ISO-8859-7",
-    "ISO-8859-8",
-    "ISO-8859-9",
-    "ISO-8859-10",
-    "ISO-8859-11",
-    "ISO-8859-13",
-    "ISO-8859-14",
-    "ISO-8859-15",
-    "ISO-8859-16",
-]:
+def _check_encoding(argstr: str) -> Encoding:
     """
     Check the charset encoding of a string.
 
@@ -177,10 +187,9 @@ def _check_encoding(
     adobe_chars = set(charset["Symbol"].values()) | set(
         charset["ZapfDingbats"].values()
     )
-    for encoding in ["ISOLatin1+"] + [f"ISO-8859-{i}" for i in range(1, 17)]:
-        if encoding == "ISO-8859-12":  # ISO-8859-12 was abandoned. Skip it.
-            continue
-        if all(c in (set(charset[encoding].values()) | adobe_chars) for c in argstr):
+    for encoding in ["ISOLatin1+"] + [f"ISO-8859-{i}" for i in range(1, 17) if i != 12]:
+        chars = set(charset[encoding].values()) | adobe_chars
+        if all(c in chars for c in argstr):
             return encoding  # type: ignore[return-value]
     # Return the "ISOLatin1+" encoding if the string contains characters from multiple
     # charset encodings or contains characters that are not in any charset encoding.
@@ -332,34 +341,13 @@ def data_kind(
     return kind  # type: ignore[return-value]
 
 
-def non_ascii_to_octal(
-    argstr: str,
-    encoding: Literal[
-        "ascii",
-        "ISOLatin1+",
-        "ISO-8859-1",
-        "ISO-8859-2",
-        "ISO-8859-3",
-        "ISO-8859-4",
-        "ISO-8859-5",
-        "ISO-8859-6",
-        "ISO-8859-7",
-        "ISO-8859-8",
-        "ISO-8859-9",
-        "ISO-8859-10",
-        "ISO-8859-11",
-        "ISO-8859-13",
-        "ISO-8859-14",
-        "ISO-8859-15",
-        "ISO-8859-16",
-    ] = "ISOLatin1+",
-) -> str:
+def non_ascii_to_octal(argstr: str, encoding: Encoding = "ISOLatin1+") -> str:
     r"""
     Translate non-ASCII characters to their corresponding octal codes.
 
     Currently, only non-ASCII characters in the Adobe ISOLatin1+, Adobe Symbol, Adobe
     ZapfDingbats, and ISO-8850-x (x can be in 1-11, 13-17) encodings are supported.
-    The Adobe Standard encoding is not supported yet.
+    The Adobe Standard+ encoding is not supported.
 
     Parameters
     ----------
@@ -488,7 +476,8 @@ def build_arg_list(  # noqa: PLR0912
     gmt_args = []
     for key, value in kwdict.items():
         if len(key) > 2:  # Raise an exception for unrecognized options
-            raise GMTInvalidInput(f"Unrecognized parameter '{key}'.")
+            msg = f"Unrecognized parameter '{key}'."
+            raise GMTInvalidInput(msg)
         if value is None or value is False:  # Exclude arguments that are None or False
             pass
         elif value is True:
@@ -498,17 +487,14 @@ def build_arg_list(  # noqa: PLR0912
         else:
             gmt_args.append(f"-{key}{value}")
 
-    # Convert non-ASCII characters (if any) in the arguments to octal codes
-    encoding = _check_encoding("".join(gmt_args))
-    if encoding != "ascii":
-        gmt_args = [non_ascii_to_octal(arg, encoding=encoding) for arg in gmt_args]
     gmt_args = sorted(gmt_args)
 
-    # Set --PS_CHAR_ENCODING=encoding if necessary
-    if encoding not in {"ascii", "ISOLatin1+"} and not (
-        confdict and "PS_CHAR_ENCODING" in confdict
-    ):
-        gmt_args.append(f"--PS_CHAR_ENCODING={encoding}")
+    # Convert non-ASCII characters (if any) in the arguments to octal codes and set
+    # --PS_CHAR_ENCODING=encoding if necessary
+    if (encoding := _check_encoding("".join(gmt_args))) != "ascii":
+        gmt_args = [non_ascii_to_octal(arg, encoding=encoding) for arg in gmt_args]
+        if not (confdict and "PS_CHAR_ENCODING" in confdict):
+            gmt_args.append(f"--PS_CHAR_ENCODING={encoding}")
 
     if confdict:
         gmt_args.extend(f"--{key}={value}" for key, value in confdict.items())
@@ -524,7 +510,8 @@ def build_arg_list(  # noqa: PLR0912
             or str(outfile) in {"", ".", ".."}
             or str(outfile).endswith(("/", "\\"))
         ):
-            raise GMTInvalidInput(f"Invalid output file name '{outfile}'.")
+            msg = f"Invalid output file name '{outfile}'."
+            raise GMTInvalidInput(msg)
         gmt_args.append(f"->{outfile}")
     return gmt_args
 
@@ -597,7 +584,7 @@ def launch_external_viewer(fname: str, waiting: float = 0):
         case "win32":  # Windows
             os.startfile(fname)  # type:ignore[attr-defined] # noqa: S606
         case _:  # Fall back to the browser if can't recognize the operating system.
-            webbrowser.open_new_tab(f"file://{fname}")
+            webbrowser.open_new_tab(f"file://{Path(fname).resolve()}")
     if waiting > 0:
         # Preview images will be deleted when a GMT modern-mode session ends, but the
         # external viewer program may take a few seconds to open the images.